Finished the record creation function.
TODO
FINISH THE RECORD INSERT FUNCTION. (add the record to the records HT, add the rec to the countries HT (insert it in the appropriate list),
insert the record to the appropriate virus Info struct (insert it in the __bloom filter__ and in the __vaccinated/not-vaccinated__ lists))
Insert records from files. Test record insert function.